Payment and some other class updation

+++ Backend/CollegeManagement/src/main/java/com/example/CollegeManagement/Controller/CourseController.java
@@ -1,34 +1,34 @@
 package com.example.CollegeManagement.Controller;
 
+import com.example.CollegeManagement.Dto.Message;
 import com.example.CollegeManagement.Entity.Course;
 import com.example.CollegeManagement.Service.CourseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.net.http.HttpResponse;
-
 @RestController
 @RequestMapping("/course")
 public class CourseController {
     @Autowired
     private CourseService courseService;
-
     @PostMapping("/register")
-    public ResponseEntity<HttpResponse> register(@RequestBody Course course) {
+    public ResponseEntity<?> register(@RequestBody Course course) {
+        if (course == null && !courseService.existsByStreamId(course.getStreamId()))
+            return ResponseEntity.badRequest().body(new Message("Course Already Exists with this name"));
         return courseService.register(course);
     }
     @PutMapping("/update/{id}")
-    public ResponseEntity<HttpResponse> update(@RequestBody Course course, @PathVariable("id") Long id) {
+    public ResponseEntity<?> update(@RequestBody Course course, @PathVariable("id") Long id) {
+        if (course == null && !courseService.existsById(id))
+            return ResponseEntity.badRequest().body(new Message("No Id available with this id"));
         return courseService.update(id, course);
     }
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<HttpResponse> delete(@PathVariable("id") Long id) {
+    public ResponseEntity<?> delete(@PathVariable("id") Long id) {
+        if (!courseService.existsById(id))
+            return ResponseEntity.badRequest().body(new Message("No Id available with this id"));
         return courseService.delete(id);
     }
-//    @GetMapping("/get/{stream}")
-//    public ResponseEntity<?> getStream(@PathVariable("stream") String stream) {
-//        return courseService.getStream(stream);
-//    }
 
 }
